+++ src/features/bidding/components/bids-list.tsx
@@ -2,9 +2,9 @@
 
 import { useState } from 'react';
 import { Button } from '@/components/ui/button';
-import { Doc } from '../../../../../../convex/_generated/dataModel';
+import { Doc } from '@convex/_generated/dataModel';
 import { usePaginatedQuery } from 'convex/react';
-import { api } from '../../../../../../convex/_generated/api';
+import { api } from '@convex/_generated/api';
 import BidCard from './bid-card';
 
 const BidsList = ({ post }: { post: Doc<'rentalposts'> }) => {
@@ -21,7 +21,7 @@ const BidsList = ({ post }: { post: Doc<'rentalposts'> }) => {
         <h2 className='text-2xl font-bold text-white'>Manage Bids</h2>
         <div className='flex gap-2'>
           <Button variant='outline'>End Bidding Early</Button>
-          {}
+          { }
           <Button>Auto-Accept Highest</Button>
         </div>
       </div>
