+++ src/features/rental/components/bidding-form.tsx
@@ -13,8 +13,8 @@ import { Slider } from '@/components/ui/slider';
 import { Button } from '@/components/ui/button';
 import { Separator } from '@/components/ui/separator';
 import { Textarea } from '@/components/ui/textarea';
-import { api } from '../../../../../../convex/_generated/api';
-import { Doc } from '../../../../../../convex/_generated/dataModel';
+import { api } from '@convex/_generated/api';
+import { Doc } from '@convex/_generated/dataModel';
 import { useSession } from 'next-auth/react';
 import z from 'zod';
 import { toast } from 'sonner';
