+++ src/features/rental/components/bidding-tab.tsx
@@ -2,7 +2,7 @@
 
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
 import { TrendingUp, Clock, User, Award, ChevronDown } from 'lucide-react';
-import { type Doc } from '../../../../../../convex/_generated/dataModel';
+import { type Doc } from '@convex/_generated/dataModel';
 import { BiddingForm } from './bidding-form';
 import { usePaginatedQuery } from 'convex/react';
 import { formatDistanceToNow } from 'date-fns';
@@ -10,9 +10,9 @@ import { Badge } from '@/components/ui/badge';
 import { Skeleton } from '@/components/ui/skeleton';
 import { Button } from '@/components/ui/button';
 import { useSession } from 'next-auth/react';
-import { api } from '../../../../../../convex/_generated/api';
+import { api } from '@convex/_generated/api';
 import { formatDuration } from '@/lib/utils';
-import { useState, useMemo } from 'react';
+import { useMemo } from 'react';
 
 type BiddingTabProps = {
   rentalPost: Doc<'rentalposts'>;
@@ -62,22 +62,20 @@ function BidItem({ bid, rank, isCurrentUser }: { bid: any; rank: number; isCurre
 
   return (
     <div
-      className={`p-4 rounded-lg border ${
-        isCurrentUser ? 'bg-blue-900/20 border-blue-700/50' : 'bg-slate-800/30 border-slate-700/50'
-      }`}>
+      className={`p-4 rounded-lg border ${isCurrentUser ? 'bg-blue-900/20 border-blue-700/50' : 'bg-slate-800/30 border-slate-700/50'
+        }`}>
       <div className='flex justify-between items-start mb-2'>
         <div className='flex items-center space-x-2'>
           {/* Ranking badge based on total value */}
           <div
-            className={`w-6 h-6 rounded-full flex items-center justify-center text-xs font-bold ${
-              rank === 1
-                ? 'bg-yellow-500 text-black'
-                : rank === 2
-                  ? 'bg-gray-400 text-black'
-                  : rank === 3
-                    ? 'bg-amber-700 text-white'
-                    : 'bg-slate-700 text-slate-300'
-            }`}>
+            className={`w-6 h-6 rounded-full flex items-center justify-center text-xs font-bold ${rank === 1
+              ? 'bg-yellow-500 text-black'
+              : rank === 2
+                ? 'bg-gray-400 text-black'
+                : rank === 3
+                  ? 'bg-amber-700 text-white'
+                  : 'bg-slate-700 text-slate-300'
+              }`}>
             #{rank}
           </div>
