+++ src/features/rental/components/rental-header.tsx
@@ -1,6 +1,4 @@
-'use client';
-
-import { type Doc } from '../../../../../../convex/_generated/dataModel';
+import { type Doc } from '@convex/_generated/dataModel';
 
 type RentalHeaderProps = {
   rentalPost: Doc<'rentalposts'>;
