+++ src/features/rental/components/rental-post-card.tsx
@@ -6,11 +6,10 @@ import { Badge } from '@/components/ui/badge';
 import { Heart, Share2, Trash2Icon } from 'lucide-react';
 import { useState, useRef, useEffect } from 'react';
 import { gsap } from 'gsap';
-import { Card3D } from '../card-3d';
+import { Card3D } from '../../../components/shared/card-3d';
 import { cn } from '@/lib/utils';
-import LendrButton from '../lendr-btn';
+import LendrButton from '../../../components/shared/lendr-btn';
 import { Doc, Id } from '../../../../convex/_generated/dataModel';
-import { usePathname } from 'next/navigation';
 import { useSession } from 'next-auth/react';
 import { Button } from '@/components/ui/button';
 import { useMutation } from 'convex/react';
